Lattigo: Refactor pass name and Support debug
