+++ lib/Pipelines/ArithmeticPipelineRegistration.h
@@ -36,7 +36,7 @@ struct MlirToRLWEPipelineOptions
       llvm::cl::init(false)};
 };
 
-struct OpenfheOptions : public PassPipelineOptions<OpenfheOptions> {
+struct BackendOptions : public PassPipelineOptions<BackendOptions> {
   PassOptions::Option<std::string> entryFunction{
       *this, "entry-function", llvm::cl::desc("Entry function"),
       llvm::cl::init("main")};
@@ -47,31 +47,11 @@ struct OpenfheOptions : public PassPipelineOptions<OpenfheOptions> {
       llvm::cl::init(false)};
 };
 
-struct LattigoOptions : public PassPipelineOptions<LattigoOptions> {
-  PassOptions::Option<int> ciphertextDegree{
-      *this, "ciphertext-degree",
-      llvm::cl::desc("The degree of the polynomials to use for ciphertexts; "
-                     "equivalently, the number of messages that can be packed "
-                     "into a single ciphertext."),
-      llvm::cl::init(1024)};
-  PassOptions::Option<std::string> entryFunction{
-      *this, "entry-function", llvm::cl::desc("Entry function"),
-      llvm::cl::init("main")};
-  PassOptions::Option<bool> modulusSwitchBeforeFirstMul{
-      *this, "modulus-switch-before-first-mul",
-      llvm::cl::desc("Modulus switching right before the first multiplication "
-                     "(default to false)"),
-      llvm::cl::init(false)};
-};
-
 using RLWEPipelineBuilder =
     std::function<void(OpPassManager &, const MlirToRLWEPipelineOptions &)>;
 
 using BackendPipelineBuilder =
-    std::function<void(OpPassManager &, const OpenfheOptions &)>;
-
-using LattigoPipelineBuilder =
-    std::function<void(OpPassManager &, const LattigoOptions &)>;
+    std::function<void(OpPassManager &, const BackendOptions &)>;
 
 void mlirToRLWEPipeline(OpPassManager &pm,
                         const MlirToRLWEPipelineOptions &options,
@@ -83,7 +63,7 @@ RLWEPipelineBuilder mlirToRLWEPipelineBuilder(RLWEScheme scheme);
 
 BackendPipelineBuilder toOpenFhePipelineBuilder();
 
-LattigoPipelineBuilder mlirToLattigoRLWEPipelineBuilder(RLWEScheme scheme);
+BackendPipelineBuilder toLattigoPipelineBuilder();
 
 }  // namespace mlir::heir
