+++ lib/Utils/BUILD
@@ -9,6 +9,7 @@ cc_library(
     hdrs = ["Utils.h"],
     deps = [
         "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:FuncDialect",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:Support",
     ],
